requirements.txt
flask-cors==3.0.10
gunicorn==20.1.0
openai==1.58.1
Flask==3.0.0
httpx==0.27.2
Authlib==1.3.1
PyJWT==2.9.0
